laurel_sprout: Use SYSTEM_OTHER_ODEX for Dexpreopt
To save on the system partition size, the preopted files can be installed in the unused second system partition.
They are then copied to the data partition on first boot.
Also optimised a bunch of dexpreopt flags.
This has been tested to have a really significant boost in the first boot time, making it nearly same as any other boot.

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>